[init] A place for random junk
I was working on some 1Password automation and wanted to keep track of a
one-liner I wrote to bulk-edit tags for a bunch of items.

After realizing the number of hours wasted on writing one-liners and
one-offs like these that have gone untracked and forgotten, I finally
decided to create a repo to store them in.

I hope this helps someone in the future!

Signed-off-by: Carlos Nunez <13461447+carlosonunez@users.noreply.github.com>